Conflicts:
	src/main/webapp/js/uploadBoard.js

+++ src/main/java/com/indiScene/audioProcessing/KOSTAAudio.java
@@ -27,6 +27,7 @@ import be.tarsos.dsp.io.jvm.WaveformWriter;
  */
 
 public class KOSTAAudio {
+	private String dir = "C:/SPB_Data/git/IndiScene/src/main/webapp/resources/MergeMusic/";
 	/**
 	 * @name : mergeAudio
 	 * @date : 2015. 7. 6.
@@ -37,6 +38,7 @@ public class KOSTAAudio {
 		byte[] outArray = null;
 		AudioInputStream ais1=null;
 		AudioInputStream ais2=null;
+		long sync =0;
 		try {
 			ais1= AudioSystem.getAudioInputStream(new File(filePath1));
 			ais2= AudioSystem.getAudioInputStream(new File(filePath2));
@@ -55,35 +57,45 @@ public class KOSTAAudio {
 			while (count !=-1){
 				count=ais2.read(buffer2,0,numBytes2);
 			}
-			outArray = new byte[Math.max(numBytes1, numBytes2)];
+			
+			sync = (long)(syncSecond*ais1.getFormat().getSampleRate()*ais1.getFormat().getChannels()*ais1.getFormat().getSampleSizeInBits()/8);
+			outArray = new byte[(int)Math.max(sync>=0?buffer1.length:buffer1.length-sync, sync<0?buffer2.length:buffer2.length+sync)];
 			System.out.println(buffer1.length +"\n" + buffer2.length+"\n"+outArray.length);
 			
-			long sync = (long)(syncSecond*ais1.getFormat().getSampleRate()*ais1.getFormat().getChannels()*ais1.getFormat().getSampleSizeInBits()/8);
-			
-			for(int i = 0 ; i < Math.max(buffer1.length, buffer2.length); i++){
-				if(i < buffer1.length && i < buffer2.length){
+			for(int i = 0 ; i < outArray.length; i++){
+				if(i < (sync>=0?buffer1.length:buffer1.length-sync) && i < (sync<0?buffer2.length:buffer2.length+sync)){
 					if(i <45){
-						outArray[i] = buffer2[i];
+						System.out.println("buffer1["+i+"] : " + buffer1[i] +"\tbuffer2["+i+"] : "+buffer2[i]);
+						outArray[i] = buffer1[i];
 					}else{
 						if(sync>=0){
 							if(i>=sync)
-								outArray[i] = (byte) ((int)(buffer1[i]*0.5 + buffer2[i]*0.5));
+								outArray[i] = (byte) ((int)(buffer1[i]*0.5 + buffer2[(int) (i-sync)]*0.5));
 							else
 								outArray[i] = buffer1[i];
 						}else{
 							if(i>=(sync*(-1)))
-								outArray[i] = (byte) ((int)(buffer1[i]*0.5 + buffer2[i]*0.5));
+								outArray[i] = (byte) ((int)(buffer1[(int) (i+sync)]*0.5 + buffer2[i]*0.5));
 							else
 								outArray[i] = buffer2[i];
 						}
 					}
 				}else{
-					outArray[i] = buffer1.length>buffer2.length?buffer1[i]:buffer2[i];
+					if(sync>=0){
+						outArray[i] = buffer1.length > buffer2.length+sync ? buffer1[i]:buffer2[(int) (i-sync)];
+					}else{
+						outArray[i] = buffer1.length - sync> buffer2.length ? buffer1[(int) (i+sync)]:buffer2[i];
+					}
 				}
 			}
 		} catch (UnsupportedAudioFileException e) {
 			e.printStackTrace();
 		} catch (IOException e) {
+			e.printStackTrace();
+		}catch (ArrayIndexOutOfBoundsException e){
+			System.out.println("outArray length : "+outArray.length);
+			System.out.println("sync : "+sync);
+			
 			e.printStackTrace();
 		}
 		
@@ -99,8 +111,14 @@ public class KOSTAAudio {
 		} catch (LineUnavailableException e) {
 			e.printStackTrace();
 		}finally{
-			File tempFile =new File(mergeFile);
-			tempFile.delete();
+			// 서버가 액세스하고있어서 삭제가 안된 파일을 삭제한다.
+			File directory =new File(dir);
+			File[] fileList = directory.listFiles();
+			for(File tempFile : fileList){
+				if(tempFile.isFile() && !tempFile.getName().substring(tempFile.getName().length()-4).equals(".wav")){
+					tempFile.delete();
+				}
+			}
 		}
 		
 		return outputFilePath;
@@ -114,10 +132,11 @@ public class KOSTAAudio {
 	 */
 	private String makeFile(byte[] byteBuffer, AudioInputStream ais, String userId) throws IOException{
 		String timeName = userId+"_"+System.currentTimeMillis();
+//		String outputFilePath = "C:/KMS_MavenSpring/apache-tomcat-7.0.59/wtpwebapps/IndiScene/resources/MergeMusic/"+timeName;
 		String outputFilePath = "C:/SPB_Data/git/IndiScene/src/main/webapp/resources/MergeMusic/"+timeName;
 		
 		ByteArrayInputStream bais = new ByteArrayInputStream(byteBuffer);
-		AudioSystem.write(new AudioInputStream(bais,ais.getFormat(),ais.getFrameLength()), AudioFileFormat.Type.WAVE, new File(outputFilePath));
+		AudioSystem.write(new AudioInputStream(bais,ais.getFormat(),byteBuffer.length), AudioFileFormat.Type.WAVE, new File(outputFilePath));
 		
 		return outputFilePath;
 	}
@@ -135,8 +154,8 @@ public class KOSTAAudio {
 		float[] floatBuffer = convertToFloat(byteBuffer);
 
 		String timeName = userId+"_"+System.currentTimeMillis()+"_Collabo.wav";
+//		String outputFilePath = "C:/Users/KOSTA/git/IndiScene/src/main/webapp/resources/MergeMusic/"+timeName;
 		String outputFilePath = "C:/SPB_Data/git/IndiScene/src/main/webapp/resources/MergeMusic/"+timeName;
-		
 		
 		ByteArrayInputStream bais = new ByteArrayInputStream(byteBuffer);
 		AudioInputStream inputStream = new AudioInputStream(bais, JVMAudioInputStream.toAudioFormat(format),floatBuffer.length);
